+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed a typo in `docker/template_container/Dockerfile_homelette_template`.
 - Fixed the suggested command to build the template container.
 - Fixed access to extension submodule.
+- Fixed routine tags for complex modelling routines.
 
 ## [1.1] - 2021-09-08
 ### Added